Add EntriesCache collection.
It's the same data as entries in Kits; as far as I can tell there's no way to get that sorted directly in graphql. Another option is to query for all entries and transform it once queried, but I'd need to query all entries every time I want some entries. Not a problem at my scale, but hey, the principle counts. :D